fix: setting tabs
