+++ manifest.json
@@ -1,7 +1,7 @@
 {
 	"id": "mathlinks",
 	"name": "MathLinks",
-	"version": "0.4.3",
+	"version": "0.4.4",
 	"minAppVersion": "1.3.5",
 	"description": "Render MathJax in your links",
 	"author": "zhaoshenzhai",
