+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "obsidian-mathlinks",
-    "version": "0.4.3",
+    "version": "0.4.4",
     "description": "Render MathJax in your links",
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
