+++ src/main.ts
@@ -1,8 +1,9 @@
 import { FileView, MarkdownEditView, MarkdownView, Plugin, WorkspaceLeaf, loadMathJax } from "obsidian";
-import { MathLinksSettings, MathLinksSettingTab, DEFAULT_SETTINGS } from "./settings";
-import { MathLinksAPIAccount, informChange } from "./api";
-import { generateMathLinks } from "./links";
-import { buildLivePreview } from "./preview";
+import { MathLinksSettings, DEFAULT_SETTINGS } from "./settings/settings";
+import { MathLinksSettingTab } from "./settings/tab"
+import { MathLinksAPIAccount, informChange } from "./api/api";
+import { generateMathLinks } from "./links/reading";
+import { buildLivePreview } from "./links/preview";
 import { isValid } from "./utils";
 
 export default class MathLinks extends Plugin {
@@ -47,12 +48,7 @@ export default class MathLinks extends Plugin {
         let account = this.apiAccounts.find((account) => account.manifest.id == userPlugin.manifest.id);
         if (account) return account;
 
-        account = new MathLinksAPIAccount(
-            this, 
-            userPlugin.manifest, 
-            DEFAULT_SETTINGS.blockPrefix, 
-            DEFAULT_SETTINGS.enableFileNameBlockLinks
-        );
+        account = new MathLinksAPIAccount(this, userPlugin.manifest, DEFAULT_SETTINGS.blockPrefix, DEFAULT_SETTINGS.enableFileNameBlockLinks);
         this.apiAccounts.push(account);
         return account;
     }
