+++ test_vault/.obsidian/plugins/math-booster/manifest.json
@@ -1,1 +1,1 @@
-{"id":"math-booster","name":"Math Booster","version":"0.3.0","minAppVersion":"1.3.5","description":"Introduce LaTeX-like mathematic note-taking features into Obsidian","author":"Ryota Ushio","authorUrl":"https://github.com/RyotaUshio","fundingUrl":"https://www.buymeacoffee.com/ryotaushio","isDesktopOnly":false}
+{"id":"math-booster","name":"Math Booster","version":"0.5.3","minAppVersion":"1.3.5","description":"Introduce LaTeX-like mathematic note-taking features into Obsidian","author":"Ryota Ushio","authorUrl":"https://github.com/RyotaUshio","fundingUrl":"https://www.buymeacoffee.com/ryotaushio","isDesktopOnly":false}
