+++ test_vault/.obsidian/plugins/math-booster/styles.css
@@ -1,12 +1,34 @@
-
 /* This file is auto-generated. Do not edit it. */
+
+.math-booster-begin-proof {
+    font-style: italic;
+    padding-right: 10px;
+}
+
+.math-booster-end-proof {
+    float: right;
+}
+
+.math-booster-add-profile {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.math-booster-add-profile > input {
+    width: 200px;
+}
+
+.math-booster-profile-button-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    gap: var(--size-4-2);
+    padding: var(--size-4-2);
+}
 
 .math-booster-preview {
     cursor: text;
-}
-
-.cm-embed-block {
-    position: relative;
 }
 
 .cm-embed-block .math-booster-preview-edit-button {
@@ -18,23 +40,23 @@
     display: flex;
     border-radius: var(--radius-s);
 }           
-
+                
 .cm-embed-block:hover .math-booster-preview-edit-button {
     transition: 0s;
     opacity: 1
 }
-
+                
 /* Unlike in callouts, Obsidian natively renders MathJax in blockquotes. But it is rather buggy 
 (see https://forum.obsidian.md/t/live-preview-support-math-block-in-quotes/32564/2), 
 so we need to replace it with this plugin's editor extension. */
 .cm-line .math.math-block.cm-embed-block:has(> mjx-container.MathJax:not(.math-booster-preview) > mjx-math[display="true"]) {
     display: none;
 }
-
+                
 .math-callout {
     position: relative;
 }
-
+                
 .math-callout-setting-button {
     padding-bottom: var(--size-2-2);
     padding-right: var(--size-2-3);
@@ -47,6 +69,77 @@ so we need to replace it with this plugin's editor extension. */
 .math-callout:hover .math-callout-setting-button {
     transition: 0s;
     opacity: 1
+}
+
+.math-callout-font-family-inherit {
+    font-family: inherit !important;
+}
+
+.math-booster-title-form,
+.math-booster-label-form {
+    width: 300px;
+}
+
+.math-callout-vivid.math-callout {
+    --callout-color: 238, 15, 149;
+    border-top: none;
+    border-bottom: none;
+    border-left: 5px solid rgb(var(--callout-color));
+    border-right: none;
+    border-radius: 0px;
+    box-shadow: none;
+    padding: 0px;
+    font-family: CMU Serif;
+}
+
+.math-callout-vivid.math-callout .callout-title {
+    padding: 6px;
+    padding-left: 12px;
+}
+
+.math-callout-vivid.math-callout .callout-icon {
+    display: none;
+}
+
+.math-callout-vivid.math-callout .callout-title-inner {
+    font-family: Inter;
+    font-weight: 400;
+    color: rgb(var(--callout-color));
+}
+
+.math-callout-vivid .math-callout-subtitle {
+    font-weight: 200;
+}
+
+.math-callout-vivid.math-callout .callout-content {
+    background-color: var(--background-primary);
+    padding: 1px 20px 2px 20px;
+}
+
+.math-callout-plain.math-callout {
+    --callout-color: var(--text-normal);
+    font-family: CMU Serif;
+    background-color: rgb(0, 0, 0, 0);
+    padding-left: 0;
+    padding-right: 0;
+    border: none;
+    box-shadow: none;
+}
+
+.math-callout-plain.math-callout .callout-icon {
+    display: none;
+}
+
+.math-callout-plain.math-callout .callout-title-inner {
+    font-style: bold;
+}
+
+.math-callout-plain .math-callout-subtitle {
+    font-weight: 400;
+}
+
+.math-callout-plain.math-callout-en .callout-content {
+    font-style: italic;
 }
 
 .math-callout-mathwiki.math-callout {
@@ -86,60 +179,7 @@ so we need to replace it with this plugin's editor extension. */
     /* Font Awesome: anchor */
     --callout-icon: '<svg aria-hidden="true" focusable="false" data-prefix="fas" data-icon="anchor" class="svg-inline--fa fa-anchor fa-w-18" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512"><path fill="currentColor" d="M12.971 352h32.394C67.172 454.735 181.944 512 288 512c106.229 0 220.853-57.38 242.635-160h32.394c10.691 0 16.045-12.926 8.485-20.485l-67.029-67.029c-4.686-4.686-12.284-4.686-16.971 0l-67.029 67.029c-7.56 7.56-2.206 20.485 8.485 20.485h35.146c-20.29 54.317-84.963 86.588-144.117 94.015V256h52c6.627 0 12-5.373 12-12v-40c0-6.627-5.373-12-12-12h-52v-5.47c37.281-13.178 63.995-48.725 64-90.518C384.005 43.772 341.605.738 289.37.01 235.723-.739 192 42.525 192 96c0 41.798 26.716 77.35 64 90.53V192h-52c-6.627 0-12 5.373-12 12v40c0 6.627 5.373 12 12 12h52v190.015c-58.936-7.399-123.82-39.679-144.117-94.015h35.146c10.691 0 16.045-12.926 8.485-20.485l-67.029-67.029c-4.686-4.686-12.284-4.686-16.971 0L4.485 331.515C-3.074 339.074 2.28 352 12.971 352zM288 64c17.645 0 32 14.355 32 32s-14.355 32-32 32-32-14.355-32-32 14.355-32 32-32z"></path></svg>';
 }
-.math-callout-vivid.math-callout {
-    --callout-color: 238, 15, 149;
-    border-left: 5px solid rgb(var(--callout-color));
-    border-radius: 0px;
-    padding: 0px;
-    font-family: CMU Serif;
-}
 
-.math-callout-vivid.math-callout .callout-title {
-    padding: 6px;
-    padding-left: 12px;
-}
-
-.math-callout-vivid.math-callout .callout-icon {
-    display: none;
-}
-
-.math-callout-vivid.math-callout .callout-title-inner {
-    font-family: Inter;
-    font-weight: 400;
-    color: rgb(var(--callout-color));
-}
-
-.math-callout-vivid .math-callout-subtitle {
-    font-weight: 200;
-}
-
-.math-callout-vivid.math-callout .callout-content {
-    background-color: var(--background-primary);
-    padding: 1px 20px 2px 20px;
-}
-.math-callout-plain.math-callout {
-    --callout-color: var(--text-normal);
-    font-family: CMU Serif;
-    background-color: rgb(0, 0, 0, 0);
-    padding-left: 0;
-    padding-right: 0;
-}
-
-.math-callout-plain.math-callout .callout-icon {
-    display: none;
-}
-
-.math-callout-plain.math-callout .callout-title-inner {
-    font-style: bold;
-}
-
-.math-callout-plain .math-callout-subtitle {
-    font-weight: 400;
-}
-
-.math-callout-plain.math-callout-en .callout-content {
-    font-style: italic;
-}
 .math-callout-framed.math-callout {
     --callout-color: var(--text-normal);
     font-family: Times;
@@ -164,6 +204,3 @@ so we need to replace it with this plugin's editor extension. */
     font-style: italic;
 }
 
-.font-family-inherit {
-    font-family: inherit !important;
-}
