+++ versions.json
@@ -23,5 +23,6 @@
     "0.4.0": "1.3.5",
     "0.4.1": "1.3.5",
     "0.4.2": "1.3.5",
-    "0.4.3": "1.3.5"
+    "0.4.3": "1.3.5",
+    "0.4.4": "1.3.5"
 }
